Move resource dynamics to libraries/chain/util #2786
