+++ libraries/chain/include/steem/chain/util/rd_dynamics.hpp
@@ -6,7 +6,7 @@
 // Data structures and functions for resource dynamics "in-flight" (i.e. in the absence of user (witness) tweaking)
 // This header includes RESOURCES functionality only, no RC (price / price curve).
 
-namespace steem { namespace plugins { namespace rc {
+namespace steem { namespace chain { namespace util {
 
 struct rd_decay_params
 {
@@ -34,12 +34,12 @@ int64_t rd_compute_pool_decay(
 
 } } }
 
-FC_REFLECT( steem::plugins::rc::rd_decay_params,
+FC_REFLECT( steem::chain::util::rd_decay_params,
    (decay_per_time_unit)
    (decay_per_time_unit_denom_shift)
    )
 
-FC_REFLECT( steem::plugins::rc::rd_dynamics_params,
+FC_REFLECT( steem::chain::util::rd_dynamics_params,
    (resource_unit)
    (budget_per_time_unit)
    (pool_eq)
